Add wrapper for pthread_kill() to check if thread exists before using kill
You can enable this extra check to be sure that we use pthread_kill()
only with valid pthread id USE_LOCKMGR_SAFEKILL in version.h.

It looks like that some implementation, like the linux one, segfaults
when you are sending a signal to a non existing thread. This problem can
be fixed by using non detached thread, or by checking if the thread is still
present before sending the signal.