Update AC_INIT use to make configure handle docdir properly
This patch allows AC_INIT to properly set the PACKAGE variables in the
Makefiles.  In turn, this allows docdir use to work correctly.  With
the new behaviour, things placed in docdir go to share/doc/bacula.
Previously the would end up in share/doc.

We move the detection of the source directory to a separate
AC_CONFIG_SRCDIR macro call.

Signed-off-by: Ben Walton <bwalton@artsci.utoronto.ca>