Add additional ACL flags for restore.
When we now encounter a restore error for an ACL due to the
fact ACL support is not enabled on a filesystem we clear
the BACL_FLAG_RESTORE_NATIVE flag so we don't try any other
restores of ACLs on the same filesystem. This saves us a
lot of overhead while we know if one fails due to no support
all the others will too. We do count the errors so the restore
report will show how many ACLs couldn't be restored. This same
algoritm is already implemented for saving ACLs.