Fix detection of batch insert enabler functions.
-R is not supported when not run through libtool. For linking the
database backend this is no problem as that is done with libtool
but for detection of things like PQisthreadsafe, PQputCopyData etc
from configure this doesn't work. So we work around that problem
using a special LFLAGS variable which is only used in configure
and which is not used outside configure.