Move batch insert detection into db.m4
Moved the batch insert detection logic into the db.m4
macros so they only run when an enable command is set.
This makes configure.in much cleaner. We also keep
a list of backends which have batch insert enabled
now that we allow multiple configured backends the
simple yes or no doesn't cut it anymore so we
output all backends that have batch insert enabled.
Also fixed some problems in detecting the batch file
enablers by splitting the LDFLAGS and LIBS for
configure. So now a postgresql or mysql in a non
standard place should work better.