Enhance mountcache with rescan option after interval.
The mountcache shouldn't be a static cache that lives forever
in the filed. Mountpoint come and go on modern systems so
on a failure to lookup a mountpoint in the cache we trigger
a rescan. We also rescan the cache on lookup when more
the 30 minutes have past since the last rescan of the data.
This way we should have less stale data. Having old data
is not so much a problem but missing new data is this new
rules should make this cache somewhat robust.