Change locking scheme of the mountpoint cache.
We now lock the mount cache from the moment a caller enters
find_mntent_mapping. This way we know for sure no one can
be tampering our structures until we finish. As flush_mntent_cache
is not called from the outside and now is a function which needs
the lock hold we remove it as an external interface.