Add support for soname setting in libtool.
We allow to specify the SONAME in a library so we can create
a libbaccats-<backend>-<version>.so which has an internal
soname of libbaccats-<version>.so This way each catalog
backend is a direct replacement of the dummy libbaccats
library and copying and linking works without the ldconfig
trying to be to smart. When no symlink exists ldconfig
will create one (it might point to the wrong one however)
but if you only install one it cannot go wrong. Also a
nice side effect is that if we create the symlink ldconfig
now doesn't replace it with a new one as it thinks its a
stale or wrong link.

install-default-backend is updated to create the symlink
in the same way that ldconfig does and point to the
libbaccats-<backend>.so which points to
libbaccats-<backend>-<version>.so