Fix mysql config for older versions of mysql+config.
On older versions of mysql_config the newly used
flags are not supported. So we test if the new flag
is supported and otherwise fallback to a kludge with
some sed strings to get the info we want from the
older mysql_config options.

This fixes problems on :

Centos 6
RHEL 6
SLES 11sp1
SLES 11sp2
OpenSuse 11.4