Fix bug #1893
Only try to close an database transaction when the database
is opened. Only destroy an rwl_lock when it is initialized.
As dbcheck only calls db_init_database and then db_close_database
the database is never opened and we try to end an transaction
on a non-opened database, the rwl_destroy will also give
an error (which is ignored) as the rwl_lock is only initialized
on opening of the database.