Fix bug #1859
SQL syntax error in UPDATE Counters and INSERT INTO Counters using
"Counters.MinValue" and "Counters.MaxValue"

Create specific queries in sql_cmds.c for the two queries in the
code that select and update counter records in the database.

All default queries in the sql_cmds.c are static from now on as
we don't want to polute the namespace with data that is only used
within sql_cmds.c to initialize the backend specific sql tables.

Removed all trailing spaces in sql_cmds.c