Reset bsmtp to only send to IPv4 mailhosts.
As part of the support for getaddrinfo which replaces the obsolete
gethostbyname interface in POSIX we can now connect to IPv6 mailhosts.
As it seems on a lot of hosts localhost is defined first as IPv6
address and then as IPv4 address. As a result of this bsmtp tries
to connect to a local mailhost using the ipv6 protocol but most
people forget to allow relay rights for IPv6 local SMTP traffic
and as such the mail is dropped. So for now we revert the behaviour
of bsmtp to only connect to IPv4 mailhosts unless the -a option
is given which will query the name service for any suitable protocol.