Fix bug #1975 new label ignored when first is duplicate.
When labelling a new volume, if an existing volume name is entered,
subsequent entries are ignored and the original one is attempted
each time.

After the first call to db_get_media_record the MediaId of the media
record is set and as such it will keep searching on MediaId instead
of using the new entered named. So clear the MediaId and set the
VolumeName and the call db_get_media_record so it will actually search
for the new name and not for the already existing mediaid of the
duplicate name entered before.